Convert all double to float for ESP32 FPU optimization
Replace all double-precision math with single-precision float throughout
the motion control system. The ESP32's FPU only supports 32-bit float;
double operations are software-emulated (~3x slower).

Changes:
- SCurve: Profile struct now uses float, removed ProfileF (redundant)
- MotionPlanner: All parameters and calculations use float
- PolarControl: MotionSettings and position handling use float
- Test harness: Updated to use float throughout

For motor control with ~0.001 precision requirements, float's ~7 digits
of precision is more than sufficient. All 24 pattern tests pass.

+++ lib/PolarControl/src/PolarControl.cpp
@@ -80,7 +80,7 @@ void PolarControl::begin() {
 }
 
 void PolarControl::updateSpeedSettings() {
-    double speedFactor = m_speed / 10.0;
+    float speedFactor = m_speed / 10.0f;
     m_planner.setSpeedMultiplier(speedFactor);
 }
 
@@ -265,7 +265,7 @@ bool PolarControl::loadAndRunFile(String filePath) {
     return loadAndRunFile(filePath, R_MAX);
 }
 
-bool PolarControl::loadAndRunFile(String filePath, double maxRho) {
+bool PolarControl::loadAndRunFile(String filePath, float maxRho) {
     xSemaphoreTake(m_mutex, portMAX_DELAY);
 
     if (m_state != IDLE) {
@@ -361,14 +361,14 @@ PolarControl::State_t PolarControl::getState() {
 }
 
 PolarCord_t PolarControl::getCurrentPosition() const {
-    double theta, rho;
+    float theta, rho;
     // Note: casting away const for the planner call - it's thread-safe
     const_cast<MotionPlanner&>(m_planner).getCurrentPosition(theta, rho);
     return {theta, rho};
 }
 
 PolarCord_t PolarControl::getActualPosition() {
-    double theta, rho;
+    float theta, rho;
     m_planner.getCurrentPosition(theta, rho);
     return {theta, rho};
 }
@@ -855,9 +855,9 @@ void PolarControl::testThetaContinuous() {
     if (m_state != IDLE) return;
     LOG("Starting theta continuous test...\r\n");
 
-    double currentTheta, currentRho;
+    float currentTheta, currentRho;
     m_planner.getCurrentPosition(currentTheta, currentRho);
-    double testRho = (currentRho > 50 && currentRho < R_MAX - 50) ? currentRho : R_MAX / 2;
+    float testRho = (currentRho > 50.0f && currentRho < R_MAX - 50.0f) ? currentRho : R_MAX / 2.0f;
 
     resetTheta();
     start(std_patch::make_unique<TestThetaContinuousGen>(testRho));
@@ -867,9 +867,9 @@ void PolarControl::testThetaStress() {
     if (m_state != IDLE) return;
     LOG("Starting theta stress test...\r\n");
 
-    double currentTheta, currentRho;
+    float currentTheta, currentRho;
     m_planner.getCurrentPosition(currentTheta, currentRho);
-    double testRho = (currentRho > 50 && currentRho < R_MAX - 50) ? currentRho : R_MAX / 2;
+    float testRho = (currentRho > 50.0f && currentRho < R_MAX - 50.0f) ? currentRho : R_MAX / 2.0f;
 
     resetTheta();
     start(std_patch::make_unique<TestThetaStressGen>(testRho));
